fix: make deck api more hook like
